# tsdb: reuse iterators to save garbage [INTERFACE CHANGE]