--- a/promql/value.go
+++ b/promql/value.go
@@ -364,6 +364,10 @@ func (ss *StorageSeries) Labels() labels.Labels {
 
 // Iterator returns a new iterator of the data of the series.
 func (ss *StorageSeries) Iterator(it chunkenc.Iterator) chunkenc.Iterator {
+	if ssi, ok := it.(*storageSeriesIterator); ok {
+		ssi.reset(ss.series)
+		return ssi
+	}
 	return newStorageSeriesIterator(ss.series)
 }
 
@@ -377,6 +381,11 @@ func newStorageSeriesIterator(series Series) *storageSeriesIterator {
 		points: series.Points,
 		curr:   -1,
 	}
+}
+
+func (ssi *storageSeriesIterator) reset(series Series) {
+	ssi.points = series.Points
+	ssi.curr = -1
 }
 
 func (ssi *storageSeriesIterator) Seek(t int64) chunkenc.ValueType {