Patterned after `Chunk.Iterator()`: pass the old iterator in so it can be re-used to avoid allocating a new object.

The PR is split into 5 commits which I think make review easier.

This requires a change to `chunks.Iterator` and `chunkenc.Iterator` interfaces.  I think this breaking change is worth it because garbage comes down by half on some benchmarks and it's very easy to match the new interface by passing `nil` or ignoring the new parameter.

In `tsdb/querier.go` we turn two levels of function closure into a single object that
 holds the required data.

In `NewListSeries` we hoist the conversion to an interface value out so it only allocates once.

In `blockBaseSeriesSet.Next()` we re-use chunk metadata structs which saves a lot when scanning the index.

And a fix to `BenchmarkMergeSeriesSet()`, which was only really doing one iteration before, regardless of what N was passed in.

<details> 
<summary><b>Benchmarks:</b></summary>

```
name                          old time/op    new time/op    delta
NoMergeSeriesSet_100_100-8      99.3µs ± 0%    96.6µs ± 1%   -2.79%  (p=0.008 n=5+5)
MergeSeriesSet/1_100_100-8       101µs ± 0%      97µs ± 0%   -4.79%  (p=0.008 n=5+5)
MergeSeriesSet/10_100_100-8     10.7ms ± 1%    10.9ms ± 2%     ~     (p=0.310 n=5+5)
MergeSeriesSet/100_100_100-8     186ms ± 2%     187ms ± 1%     ~     (p=0.151 n=5+5)

name                          old alloc/op   new alloc/op   delta
NoMergeSeriesSet_100_100-8      4.90kB ± 0%    0.10kB ± 3%  -97.95%  (p=0.008 n=5+5)
MergeSeriesSet/1_100_100-8      4.91kB ± 0%    0.11kB ± 0%  -97.72%  (p=0.008 n=5+5)
MergeSeriesSet/10_100_100-8      211kB ± 0%     139kB ± 0%  -34.13%  (p=0.008 n=5+5)
MergeSeriesSet/100_100_100-8    1.86MB ± 0%    1.19MB ± 0%  -35.85%  (p=0.008 n=5+5)
```

Here I filtered to just the ones operating on a hundred or more series; the smaller ones show less benefit.
```
name                                                                                                      old time/op    new time/op    delta
RangeQuery/expr=a_hundred,steps=1-8                                                                          216µs ± 0%     201µs ± 0%   -6.68%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred,steps=10-8                                                                         238µs ± 4%     217µs ± 0%   -8.59%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred,steps=100-8                                                                        551µs ± 1%     530µs ± 0%   -3.90%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred,steps=1000-8                                                                      2.99ms ± 1%    2.89ms ± 1%   -3.54%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=1-8                                                                242µs ± 2%     224µs ± 0%   -7.23%  (p=0.016 n=5+4)
RangeQuery/expr=rate(a_hundred[1m]),steps=10-8                                                               297µs ± 0%     281µs ± 0%   -5.38%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=100-8                                                             1.02ms ± 0%    1.01ms ± 0%   -1.83%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-8                                                            7.15ms ± 1%    7.10ms ± 2%     ~     (p=0.151 n=5+5)
RangeQuery/expr=rate(a_hundred[1m]),steps=10000-8                                                           78.5ms ± 1%    78.2ms ± 1%     ~     (p=0.421 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-8                                             37.4ms ± 1%    37.0ms ± 0%   -1.06%  (p=0.016 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-8                                            89.6ms ± 0%    89.8ms ± 1%     ~     (p=0.421 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-8                                            615ms ± 1%     617ms ± 0%     ~     (p=0.151 n=5+5)
RangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-8                                           5.86s ± 0%     5.86s ± 0%     ~     (p=0.310 n=5+5)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-8                                                            27.3ms ± 1%    27.2ms ± 0%     ~     (p=0.095 n=5+5)
RangeQuery/expr=changes(a_hundred[1d]),steps=10-8                                                           36.0ms ± 1%    35.7ms ± 1%     ~     (p=0.056 n=5+5)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-8                                                           122ms ± 1%     122ms ± 0%     ~     (p=0.905 n=5+4)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-8                                                          976ms ± 1%     970ms ± 0%   -0.60%  (p=0.016 n=5+4)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-8                                                               27.6ms ± 1%    27.2ms ± 1%   -1.77%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_hundred[1d]),steps=10-8                                                              36.2ms ± 1%    36.1ms ± 1%     ~     (p=0.222 n=5+5)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-8                                                              123ms ± 3%     122ms ± 2%     ~     (p=0.310 n=5+5)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-8                                                             979ms ± 0%     983ms ± 1%     ~     (p=0.151 n=5+5)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-8                                                   26.6ms ± 0%    26.4ms ± 1%   -0.88%  (p=0.016 n=5+5)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-8                                                  30.5ms ± 1%    30.3ms ± 0%     ~     (p=0.151 n=5+5)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-8                                                 67.5ms ± 0%    67.4ms ± 1%     ~     (p=0.421 n=5+5)
RangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-8                                                 435ms ± 0%     436ms ± 1%     ~     (p=0.310 n=5+5)
RangeQuery/expr=-a_hundred,steps=1-8                                                                         237µs ± 0%     224µs ± 1%   -5.61%  (p=0.016 n=4+5)
RangeQuery/expr=-a_hundred,steps=10-8                                                                        254µs ± 1%     239µs ± 0%   -6.07%  (p=0.008 n=5+5)
RangeQuery/expr=-a_hundred,steps=100-8                                                                       579µs ± 0%     560µs ± 0%   -3.37%  (p=0.008 n=5+5)
RangeQuery/expr=-a_hundred,steps=1000-8                                                                     3.10ms ± 0%    2.99ms ± 1%   -3.62%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1-8                                                              552µs ± 0%     524µs ± 0%   -5.17%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_-_b_hundred,steps=10-8                                                             944µs ± 0%     915µs ± 0%   -3.01%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_-_b_hundred,steps=100-8                                                           5.19ms ± 0%    5.18ms ± 2%     ~     (p=0.222 n=5+5)
RangeQuery/expr=a_hundred_-_b_hundred,steps=1000-8                                                          46.6ms ± 1%    46.3ms ± 0%   -0.55%  (p=0.032 n=5+5)
RangeQuery/expr=a_hundred_-_b_hundred,steps=10000-8                                                          473ms ± 0%     473ms ± 0%     ~     (p=0.690 n=5+5)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-8                                             392µs ± 0%     374µs ± 1%   -4.58%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-8                                            555µs ± 0%     536µs ± 0%   -3.43%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-8                                          2.42ms ± 0%    2.40ms ± 1%   -1.06%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                         20.1ms ± 0%    20.0ms ± 0%   -0.78%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-8                                              423µs ± 0%     405µs ± 0%   -4.25%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-8                                             671µs ± 0%     655µs ± 0%   -2.29%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-8                                           3.41ms ± 1%    3.38ms ± 1%     ~     (p=0.095 n=5+5)
RangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                          29.9ms ± 1%    29.7ms ± 0%   -0.67%  (p=0.016 n=5+5)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1-8                                          394µs ± 1%     375µs ± 0%   -4.93%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=10-8                                         560µs ± 2%     538µs ± 0%   -4.04%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=100-8                                       2.43ms ± 0%    2.40ms ± 0%   -1.15%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1000-8                                      20.1ms ± 0%    20.0ms ± 1%   -0.84%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_and_b_hundred{l='notfound'},steps=1-8                                              223µs ± 0%     210µs ± 0%   -5.75%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_and_b_hundred{l='notfound'},steps=10-8                                             251µs ± 0%     238µs ± 0%   -5.40%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_and_b_hundred{l='notfound'},steps=100-8                                            691µs ± 0%     670µs ± 0%   -3.11%  (p=0.016 n=4+5)
RangeQuery/expr=a_hundred_and_b_hundred{l='notfound'},steps=1000-8                                          4.33ms ± 0%    4.18ms ± 0%   -3.56%  (p=0.008 n=5+5)
RangeQuery/expr=abs(a_hundred),steps=1-8                                                                     296µs ± 2%     281µs ± 0%   -5.25%  (p=0.008 n=5+5)
RangeQuery/expr=abs(a_hundred),steps=10-8                                                                    514µs ± 0%     500µs ± 0%   -2.78%  (p=0.008 n=5+5)
RangeQuery/expr=abs(a_hundred),steps=100-8                                                                  2.86ms ± 0%    2.86ms ± 0%     ~     (p=0.310 n=5+5)
RangeQuery/expr=abs(a_hundred),steps=1000-8                                                                 25.7ms ± 0%    26.1ms ± 4%   +1.61%  (p=0.016 n=5+5)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=1-8                                  337µs ± 0%     325µs ± 1%   -3.44%  (p=0.008 n=5+5)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=10-8                                 591µs ± 0%     581µs ± 0%   -1.74%  (p=0.008 n=5+5)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=100-8                               3.29ms ± 1%    3.28ms ± 0%     ~     (p=0.690 n=5+5)
RangeQuery/expr=label_replace(a_hundred,_'l2',_'$1',_'l',_'(.*)'),steps=1000-8                              29.7ms ± 0%    29.7ms ± 0%     ~     (p=0.310 n=5+5)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=1-8                                         319µs ± 1%     308µs ± 1%   -3.57%  (p=0.008 n=5+5)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=10-8                                        573µs ± 0%     561µs ± 0%   -2.23%  (p=0.008 n=5+5)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=100-8                                      3.28ms ± 1%    3.25ms ± 0%   -0.80%  (p=0.008 n=5+5)
RangeQuery/expr=label_join(a_hundred,_'l2',_'-',_'l',_'l'),steps=1000-8                                     29.8ms ± 0%    29.5ms ± 0%   -0.82%  (p=0.008 n=5+5)
RangeQuery/expr=sum(a_hundred),steps=1-8                                                                     221µs ± 1%     206µs ± 0%   -6.78%  (p=0.008 n=5+5)
RangeQuery/expr=sum(a_hundred),steps=10-8                                                                    267µs ±11%     243µs ± 0%   -8.84%  (p=0.008 n=5+5)
RangeQuery/expr=sum(a_hundred),steps=100-8                                                                   793µs ± 0%     776µs ± 1%   -2.10%  (p=0.008 n=5+5)
RangeQuery/expr=sum(a_hundred),steps=1000-8                                                                 5.43ms ± 1%    5.27ms ± 1%   -2.81%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1-8                                                        2.34ms ± 0%    2.20ms ± 0%   -6.06%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=10-8                                                       2.76ms ± 1%    2.62ms ± 0%   -4.85%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=100-8                                                      8.92ms ± 1%    8.77ms ± 3%     ~     (p=0.151 n=5+5)
RangeQuery/expr=sum_without_(l)(h_hundred),steps=1000-8                                                     74.0ms ± 1%    73.7ms ± 3%     ~     (p=0.548 n=5+5)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1-8                                                       2.40ms ± 1%    2.28ms ± 1%   -4.76%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=10-8                                                      3.13ms ± 0%    2.99ms ± 1%   -4.34%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=100-8                                                     12.3ms ± 0%    12.1ms ± 1%   -2.20%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(le)(h_hundred),steps=1000-8                                                     102ms ± 0%     101ms ± 0%   -0.79%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1-8                                                             2.39ms ± 1%    2.27ms ± 1%   -4.96%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=10-8                                                            3.13ms ± 0%    3.01ms ± 1%   -3.87%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=100-8                                                           12.4ms ± 0%    12.3ms ± 2%     ~     (p=0.151 n=5+5)
RangeQuery/expr=sum_by_(l)(h_hundred),steps=1000-8                                                           102ms ± 0%     101ms ± 0%   -0.72%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1-8                                                            2.33ms ± 1%    2.20ms ± 0%   -5.40%  (p=0.008 n=5+5)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=10-8                                                           2.79ms ± 0%    2.67ms ± 2%   -4.28%  (p=0.016 n=4+5)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=100-8                                                          9.35ms ± 0%    9.07ms ± 1%   -3.04%  (p=0.016 n=4+5)
RangeQuery/expr=sum_by_(le)(h_hundred),steps=1000-8                                                         73.1ms ± 0%    71.6ms ± 0%   -2.11%  (p=0.008 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=1-8                                                  3.96ms ± 0%    3.85ms ± 2%   -2.61%  (p=0.016 n=4+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=10-8                                                 15.0ms ± 0%    14.9ms ± 0%   -0.52%  (p=0.032 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=100-8                                                 150ms ± 1%     149ms ± 1%     ~     (p=0.310 n=5+5)
RangeQuery/expr=count_values('value',_h_hundred),steps=1000-8                                                6.89s ± 3%     6.73s ± 4%     ~     (p=0.486 n=4+4)
RangeQuery/expr=topk(1,_a_hundred),steps=1-8                                                                 222µs ± 0%     209µs ± 0%   -5.87%  (p=0.016 n=4+5)
RangeQuery/expr=topk(1,_a_hundred),steps=10-8                                                                265µs ± 3%     258µs ± 5%     ~     (p=0.095 n=5+5)
RangeQuery/expr=topk(1,_a_hundred),steps=100-8                                                               845µs ± 1%     843µs ± 1%     ~     (p=0.690 n=5+5)
RangeQuery/expr=topk(1,_a_hundred),steps=1000-8                                                             5.90ms ± 1%    5.95ms ± 2%     ~     (p=0.421 n=5+5)
RangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=1-8                                          597µs ± 0%     573µs ± 3%   -3.98%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=10-8                                        1.05ms ± 0%    1.02ms ± 0%   -2.56%  (p=0.008 n=5+5)
RangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=100-8                                       5.88ms ± 0%    5.89ms ± 0%     ~     (p=0.310 n=5+5)
RangeQuery/expr=rate(a_hundred[1m])_+_rate(b_hundred[1m]),steps=1000-8                                      52.5ms ± 1%    52.5ms ± 0%     ~     (p=0.841 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1-8                                               249µs ± 1%     233µs ± 0%   -6.39%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=10-8                                              325µs ± 1%     312µs ± 0%   -3.98%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=100-8                                            1.27ms ± 0%    1.26ms ± 0%   -0.96%  (p=0.016 n=5+4)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m])),steps=1000-8                                           9.52ms ± 0%    9.50ms ± 0%     ~     (p=0.286 n=5+4)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1-8        497µs ± 0%     469µs ± 0%   -5.79%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=10-8       654µs ± 0%     627µs ± 0%   -4.14%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=100-8     2.56ms ± 0%    2.54ms ± 0%   -1.03%  (p=0.008 n=5+5)
RangeQuery/expr=sum_without_(l)(rate(a_hundred[1m]))_/_sum_without_(l)(rate(b_hundred[1m])),steps=1000-8    19.3ms ± 0%    19.4ms ± 1%     ~     (p=0.190 n=4+5)
RangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=1-8                                      3.22ms ± 1%    3.09ms ± 0%   -3.99%  (p=0.008 n=5+5)
RangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=10-8                                     5.08ms ± 3%    4.94ms ± 0%   -2.78%  (p=0.008 n=5+5)
RangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=100-8                                    24.7ms ± 0%    24.6ms ± 0%     ~     (p=0.111 n=4+5)
RangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=1000-8                                    220ms ± 0%     221ms ± 1%     ~     (p=0.056 n=5+5)
RangeQuery/expr=a_hundred_+_on(l)_group_right_a_one,steps=1-8                                                244µs ± 0%     232µs ± 0%   -5.02%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_+_on(l)_group_right_a_one,steps=10-8                                               326µs ± 0%     314µs ± 0%   -3.63%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_+_on(l)_group_right_a_one,steps=100-8                                             1.30ms ± 0%    1.29ms ± 0%   -0.94%  (p=0.008 n=5+5)
RangeQuery/expr=a_hundred_+_on(l)_group_right_a_one,steps=1000-8                                            10.3ms ± 0%    10.2ms ± 0%   -0.51%  (p=0.008 n=5+5)

…