@codesome 2022-10-03
/prombench cancel

Some noticeable reduction in allocation. It may be a bigger difference in certain setups.